+++ src/config/globalVar.ts
@@ -9,7 +9,8 @@ export abstract class DailyTask {
   /** 是否停止投币 */
   static isStopCoin: boolean = false;
   /** 自定义up */
-  static readonly CUSTOMIZE_UP: number[] = userConfig.CUSTOMIZE_UP || [];
+  static readonly CUSTOMIZE_UP: number[] =
+    userConfig.customizeUp || userConfig.CUSTOMIZE_UP || []; // 对老版本兼容 CUSTOMIZE_UP
 
   static isRun: boolean = true;
 
@@ -18,6 +19,8 @@ export abstract class DailyTask {
   static readonly MAX_ADD_COIN_NUM = userConfig.coinTarget ?? 5;
 
   static readonly STAY_COINS = userConfig.stayCoins ?? 0;
+
+  static readonly TARGET_LEVEL = userConfig.targetLevel ?? 6;
 
   static money = 9999;
 
@@ -32,9 +35,10 @@ export abstract class DailyTask {
 
 export abstract class OSConfig {
   /**  浏览器UA */
-  static USER_AGENT: string = process.env.USER_AGENT || userConfig.USER_AGENT;
+  static USER_AGENT: string =
+    process.env.USER_AGENT || userConfig.userAgent || userConfig.USER_AGENT; // USER_AGENT 对老版本配置兼容
   /** 系统三次响应的时间戳 */
-  static COOKIE: string = process.env.BILI_COOKIE;
+  static COOKIE: string = process.env.BILI_COOKIE || userConfig.cookie;
   /**  用户id */
   static USER_ID: number = getUserId(OSConfig.COOKIE);
 }