+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Features
 
-- 更新依赖版本
-- 测试通过提交 tag 触发 action
+- 重命名一些 `config.json` 配置项
+- `config.json` 增加 `cookie` 和 `targetLevel` 配置项
+- 检测当前等级是否达到目标等级，从而确定是否取消投币